wasm: memory64 / table64 — i64 addressing → 98.78%
Implements the memory64 proposal (the *64 spec files, ~1540 failures).
Limits widen to u64 with an is_64 flag; the decoder accepts the 64-bit
limits encodings (flags 0x04-0x07). The validator types every memory
and table access by its address width — load/store/size/grow/copy/
fill/init addresses and table indices are i64 for a 64-bit memory or
table, i32 otherwise. The interpreter reads operands uniformly as u64
(an i32 zero-extends in its cell, so this is correct for both widths),
selects the result width for size/grow by is_64, and reads memarg
offsets and const-expr offsets at the right width.

Two latent bugs surfaced and fixed: the inline LEB readers'
sign-extension shifted out of range on full-width (5-byte i32 /
10-byte i64) constants, and checkBounds computed start+n directly,
which overflows u64 for near-2^64 memory64 operands — both now
range-checked without overflow.

Official spec testsuite 54248 → 55790 (96.05% → 98.78%); fails drop
2232 → 690.

+++ src/runtime/wasm/tests.zig
@@ -301,8 +301,8 @@ test "wasm decoder: decodes an import section" {
     try testing.expectEqual(@as(usize, 2), m.imports.len);
     try testing.expectEqualStrings("env", m.imports[0].module);
     try testing.expectEqualStrings("mem", m.imports[0].name);
-    try testing.expectEqual(@as(u32, 1), m.imports[0].desc.mem.limits.min);
-    try testing.expectEqual(@as(?u32, null), m.imports[0].desc.mem.limits.max);
+    try testing.expectEqual(@as(u64, 1), m.imports[0].desc.mem.limits.min);
+    try testing.expectEqual(@as(?u64, null), m.imports[0].desc.mem.limits.max);
 
     try testing.expectEqualStrings("g", m.imports[1].name);
     try testing.expectEqual(ValType.i32, m.imports[1].desc.global.val);
@@ -340,11 +340,11 @@ test "wasm decoder: decodes table and memory sections" {
 
     try testing.expectEqual(@as(usize, 1), m.tables.len);
     try testing.expectEqual(@import("types.zig").RefType.funcref, m.tables[0].elem);
-    try testing.expectEqual(@as(u32, 1), m.tables[0].limits.min);
-    try testing.expectEqual(@as(?u32, 2), m.tables[0].limits.max);
+    try testing.expectEqual(@as(u64, 1), m.tables[0].limits.min);
+    try testing.expectEqual(@as(?u64, 2), m.tables[0].limits.max);
 
     try testing.expectEqual(@as(usize, 1), m.mems.len);
-    try testing.expectEqual(@as(u32, 1), m.mems[0].limits.min);
+    try testing.expectEqual(@as(u64, 1), m.mems[0].limits.min);
 }
 
 // ── malformed inputs ────────────────────────────────────────────────
